requirements.txt
# Django REST server
django==2.1.7
gunicorn==19.9.0
# image
Pillow==6.0.0

# Timezone
pytz==2018.9

# Configuration
# https://django-environ.readthedocs.io/en/latest/
django-environ==0.4.3

# Python-PostgreSQL Database Adapter
psycopg2==2.7.7
# http://initd.org/psycopg/docs/install.html#binary-install-from-pypi
psycopg2-binary==2.7.7

# Django Rest Framework
# https://www.django-rest-framework.org/#installation
djangorestframework==3.9.2
markdown==3.0.1                 # Markdown support for the browsable API.
django-filter==2.1.0            # Filtering support
coreapi==2.3.3

# Python Data Analysis Library
pandas==0.24.2
xlrd==1.2.0

# ckeditor
# https://pypi.org/project/django-ckeditor/
django-ckeditor==5.6.1

# SendGrid for Email
sendgrid-django==4.2.0